#3: Add occurence related feature & CSV creator
This commit will add an occurence related feature which
will count the word/symbol occurence within the movie reviews. The
vectorizer is set to binary mode, so feature is either 0 or 1.

This commit will also move the previous example from main.py
to example.py file.

In addition, this commit will add a convenient method to construct
a csv file, given the author name.

closes gh-3